+++ .gitignore
@@ -7,6 +7,7 @@ yarn-debug.log*
 yarn-error.log*
 lerna-debug.log*
 
+package-lock.json
 # Runtime data
 pids
 *.pid
