+++ cloud/datasync/src/index.ts
@@ -1,68 +1,13 @@
 import * as Promise from 'bluebird';
 import * as Logger from 'bunyan';
 import * as Express from 'express';
-import SyncGlobalOptions from './syncOptions';
-
+import * as SyncApiService from './SyncApiService';
+import {FeedhenrySyncOptions, SyncDataLayerOptions, SyncDataSetOptions} from './SyncApiService';
 export * from './syncOptions';
 
 const syncAPI: any = require('fh-sync').api;
+
 const log = Logger.createLogger({ name: __filename, level: 'debug' });
-
-/**
- * Top level sync interface
- */
-export interface SyncApiService {
-
-  /**
-   * Register dataset to be monitored by sync
-   */
-  registerDataset(datasetId: string, options: SyncDataSetOptions): void;
-}
-
-export default SyncApiService;
-
-/**
- * Provide custom hash function to determine if object was changed
- */
-export interface HashFunction {
-   function(object: any): any;
-}
-
-/**
- * Required options for sync
- */
-export interface SyncDataLayerOptions {
-  mongoDbConnectionUrl: string;
-  mongoDbOptions: any;
-  redisConnectionUrl: string;
-}
-
-/**
- * Global sync options used to initialize sync
- *
- * @field datasetConfiguration - required mongodb and redis configuration
- */
-export interface FeedhenrySyncOptions {
-  datasetConfiguration: SyncDataLayerOptions;
-  globalSyncOptions?: SyncGlobalOptions;
-  globalHashFunction?: HashFunction;
-}
-
-/**
- * Complete set of options for dataset initialization
- */
-export interface SyncDataSetOptions {
-  collisionHandler?: CollisionHandler;
-  hashFunction?: HashFunction;
-}
-
-/**
- * Interface for handling collisions
- */
-export interface CollisionHandler {
-  // tslint:disable-next-line:max-line-length
-  function(datasetId: string, hash: string, timestamp: string, uid: any, pre: any, post: any, metaData: any, cb: any): void;
-}
 
 /**
  * Implementation for sync api
@@ -107,7 +52,7 @@ class FeedhenrySync implements SyncApiService {
   }
 
   private initSync(sdo: SyncDataLayerOptions) {
-    const promise = new Promise(function (resolve: () => void, reject: () => void) {
+    const promise = new Promise(function(resolve: () => void, reject: () => void) {
       const mongoDbConnectionUrl = sdo.mongoDbConnectionUrl;
       syncAPI.connect(sdo.mongoDbConnectionUrl, sdo.mongoDbOptions, sdo.redisConnectionUrl, function () {
         resolve();
